pyproject.toml
[tool.poetry]
name = "ml-ops"
version = "0.1.0"
description = ""
authors = ["Daniil Berezin"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
numpy = "^1.26.0"
catboost = "^1.2.2"
pandas = "^2.1.1"
scikit-learn = "^1.3.1"
joblib = "^1.3.2"
imblearn = "^0.0"
requests = "^2.31.0"
pre-commit = "^3.4.0"
torch = {version = "^2.1.1+cpu", source = "pytorch"}
torchvision = {version = "^0.16.1+cpu", source = "pytorch"}
torchmetrics = "^1.2.0"
dvc = "^3.30.1"
dvc-gdrive = "^2.20.0"
hydra-core = "^1.3.2"
mlflow = "^2.8.1"
tritonclient = "^2.41.0"
gevent = "^23.9.1"
geventhttpclient = "^2.0.11"


[[tool.poetry.source]]
name = "pytorch"
url = "https://download.pytorch.org/whl/cpu"
priority = "explicit"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
